requirements.txt
Django
django-admin-interface
django-archive
django-colorfield
django-crispy-forms
django-import-export
django-registration-redux
django-taggit
django-tinymce
django-versatileimagefield
docutils
isort
num2words
pdfkit
Pillow
psycopg2-binary
python-decouple
django-wkhtmltopdf
djangoql
